fixing ffmpeg issue

+++ tools/transcription.py
@@ -1,13 +1,22 @@
 import subprocess
 import tempfile
 import os
+import logging
+from sys import platform
 
-import whisper
+from whisper import Whisper
 import torch
+
+logger = logging.getLogger(__name__)
 
 
 def transcribe_audio(
-    audio_file: str, start: float, end: float, model: str = "base"
+    audio_file: str,
+    start: float,
+    end: float,
+    model: Whisper,
+    whisper_build_path: str,
+    whisper_model_path: str,
 ) -> str:
     """
     Transcribe audio to text using OpenAI's Whisper model.
@@ -23,12 +32,11 @@ def transcribe_audio(
     tmp_path = tmp_wav.name
     tmp_wav.close()
 
-    device = "cuda" if torch.cuda.is_available() else "cpu"
-
-    # Trim the audio ussing ffmpeg
+    # Trim the audio using ffmpeg
     subprocess.run(
         [
             "ffmpeg",
+            "-y",  # Overwrite output files without asking
             "-i",
             audio_file,
             "-ss",
@@ -47,11 +55,31 @@ def transcribe_audio(
         stdout=subprocess.DEVNULL,
         stderr=subprocess.DEVNULL,
     )
+    logger.debug(f"Trimmed audio saved to {tmp_path}")
 
-    # Load the Whisper model
-    model = whisper.load_model(model, device=device)
-    result = model.transcribe(tmp_path, language="en", fp16=torch.cuda.is_available())
-    transcript = result["text"]
-    os.remove(tmp_path)  # Clean up the temporary file
+    if platform == "darwin" and os.path.exists("whisper.cpp"):
+        logger.info("Using whisper.cpp for transcription on macOS")
+        cmd = [
+            whisper_build_path,
+            "-m",
+            whisper_model_path,
+            "-f",
+            tmp_path,
+            "--no-timestamps",
+        ]
+        result = subprocess.run(cmd, capture_output=True, text=True)
+        if result.returncode != 0:
+            raise Exception(f"Error in transcription: {result.stderr}")
 
-    return transcript
+        transcription = result.stdout
+        return transcription
+    else:
+        logger.info("Using Whisper Python API for transcription")
+        result = model.transcribe(
+            tmp_path, language="en", fp16=torch.cuda.is_available()
+        )
+        transcription = result["text"]
+        os.remove(tmp_path)  # Clean up the temporary file
+        logger.debug(f"Transcription result: {transcription}")
+
+        return transcription
